requirements.txt
aiogram==2.25.2
aioredis==1.3.1
redis==5.0.4
python-dotenv